fix: 修复 ARM MTE tag 导致地址解析失败和内存读写异常的问题 (#72)
- Kotlin 层：将所有 hex 地址解析从 toLong(16) 改为 toULong(16).toLong()，支持高位为 1 的地址
- Rust 层：在 read_memory_unified/write_memory_unified 入口 strip MTE tag (bits 56-63)
- 内存预览：地址指针高亮判断前先 strip MTE tag，修复带 tag 地址不变色的问题
- 跳转地址时自动 strip MTE tag

+++ app/src/main/java/moe/fuqiuluo/mamu/floating/dialog/OffsetCalculatorDialog.kt
@@ -306,7 +306,7 @@ class OffsetCalculatorDialog(
             val result = withContext(Dispatchers.IO) {
                 // 解析基址
                 val baseAddress = try {
-                    baseAddressStr.toLong(16)
+                    baseAddressStr.toULong(16).toLong()
                 } catch (e: Exception) {
                     e.printStackTrace()
                     throw ParseException("基址格式错误: $baseAddressStr")
@@ -670,7 +670,7 @@ class OffsetCalculatorDialog(
 
         val baseAddressStr = binding.inputBaseAddress.text.toString().trim()
         val baseAddress = try {
-            baseAddressStr.toLong(16)
+            baseAddressStr.toULong(16).toLong()
         } catch (e: Exception) {
             0L
         }
@@ -783,7 +783,7 @@ class OffsetCalculatorDialog(
         // 添加到历史记录
         val baseAddressStr = binding.inputBaseAddress.text.toString().trim()
         val expressionStr = binding.inputExpression.text.toString().trim()
-        val baseAddress = try { baseAddressStr.toLong(16) } catch (e: Exception) { 0L }
+        val baseAddress = try { baseAddressStr.toULong(16).toLong() } catch (e: Exception) { 0L }
         addToHistory(baseAddress, expressionStr, result.finalAddress, binding.cbHexMode.isChecked)
 
         // 使用 JumpToMemoryPreview 事件，会先切换到内存预览 Tab 再跳转
